knowledge_base/genie_metadata_generator/app/requirements.txt
streamlit>=1.39.10
databricks-sql-connector>=3.0.0
databricks-sdk>=0.18.0
pyyaml>=6.0
streamlit-ace>=0.1.1
python-dotenv>=1.0.0
requests>=2.28.0
# Navigation bar component
streamlit-navigation-bar>=3.3.0
# Lakebase PostgreSQL backend (for persistent session storage)
psycopg2-binary>=2.9.9